Moving InformationElementTemplate definition to lib
This template will be used by MT & MO.
